requirements.txt
pymongo
Pillow
boto3
python-dotenv
